+++ backend/server/routers/api.py
@@ -1,92 +1,14 @@
 from fastapi import APIRouter
-from fastapi.params import Body
 from server.database import specialisationsCOL, programsCOL, coursesCOL, archivesDB
 from fastapi.responses import JSONResponse
-from fastapi.encoders import jsonable_encoder
-from pydantic import BaseModel, create_model
 import re
-import collections
-import pickle 
 from algorithms.objects.user import User
-from typing import Dict, List 
+from server.routers.model import *
 
 router = APIRouter(
     prefix='/api',
     tags=['api'],
 )
-
-minorInFE = ['3778']
-minorInSpecialisation = ['3502', '3970']
-
-class message (BaseModel):
-    message: str
-
-class programs (BaseModel):
-    programs: dict
-
-class majors (BaseModel):
-    majors: dict
-
-class minors (BaseModel):
-    minors: dict
-
-class programCourses (BaseModel):
-    courses: dict
-
-class core (BaseModel):
-    core: dict
-
-class courseDetails (BaseModel):
-    title: str
-    code: str
-    UOC: int
-    level: int
-    description: str
-    study_level: str
-    school: str
-    campus: str
-    equivalents: dict
-    exclusions: dict
-    path_to: dict
-    terms: list
-    gen_ed: int
-    path_from: dict
-
-class course (BaseModel):
-    course: courseDetails
-
-class Structure (BaseModel):
-    structure: dict
-
-class UserData(BaseModel):
-    program: str 
-    specialisations: list
-    courses: dict
-    year: int
-
-class CourseState (BaseModel):
-    is_accurate: bool 
-    unlocked: bool
-    handbook_note: str 
-    warnings: list
-
-class CoursesState (BaseModel):
-    courses_state: Dict[str, CourseState] = {}
-
-
-class PlannerTerm (BaseModel):
-    locked: bool
-    courses: dict
-
-class PlannerData (BaseModel):
-    program: str
-    specialisations: List[str]
-    year: int
-    plan: List[List[PlannerTerm]]
-
-CONDITIONS_PATH = "algorithms/conditions.pkl"
-with open(CONDITIONS_PATH, "rb") as file:
-    CONDITIONS = pickle.load(file)
 
 def addSpecialisation(structure, code, type):
     query = {'code': code}
@@ -708,142 +630,3 @@ def unselectCourse(userData: UserData, lockedCourses: list, unselectedCourse: st
     affectedCourses =  user.unselect_course(unselectedCourse, lockedCourses)
 
     return {'affected_courses': affectedCourses}
-
-@router.post("/validateTermPlanner/", response_model=CoursesState)
-async def validateTermPlanner(plannerData: PlannerData = Body(
-    ...,
-    example={
-        "program": "3707",
-        "specialisations": ["COMPA1"],
-        "year": 1,
-        "plan": [
-            [
-                {
-                    "locked": False,
-                    "courses": {}
-                },
-                {
-                    "locked": False,
-                    "courses": {
-                        "COMP1511": [6, None],
-                        "MATH1141": [6, None],
-                        "MATH1081": [6, None],
-                    }
-                },
-                {
-                    "locked": False,
-                    "courses": {
-                        "COMP1521": [6, None],
-                        "COMP9444": [6, None],
-                    }
-                },
-                {
-                    "locked": False,
-                    "courses": {
-                        "COMP2521": [6, None],
-                        "MATH1241": [6, None],
-                        "COMP3331": [6, None]
-                    }
-                }
-            ],
-            [
-                {
-                    "locked": False,
-                    "courses": {}
-                },
-                {
-                    "locked": False,
-                    "courses": {
-                        "COMP1531": [6, None],
-                        "COMP6080": [6, None],
-                        "COMP3821": [6, None]
-                    }
-                }
-            ]
-        ]
-    }
-)):
-    """
-    Will iteratveily go through the term planner data whilst "building up" the user.
-    Starting from 1st year ST, we will create an empty user and evaluate the courses.
-    Then we will add ST courses to the user and evaluate T1. Then we will add T1
-    courses and evaluate T2. Then add T2 and evaluate T3. Then add T3 and evaluate
-    2nd year ST... and so on.
-
-    Returns the state of all the courses on the term planner
-    """
-    data = plannerData.dict()
-    emptyUserData = {
-        "program": data["program"],
-        "specialisations": data["specialisations"],
-        "year": 1, # Start off as a first year
-        "courses": {} # Start off the user with an empty year
-    }
-    user = User(emptyUserData)
-
-    # State of courses on the term planner
-    coursesState = {}
-
-    for year in data["plan"]:
-        # Go through all the years
-        for term in year:
-            # Go through all the terms. Do not bother evaluating locked terms
-            if term["locked"]:
-                for course in term["courses"]:
-                    coursesState[course] = {
-                        "is_accurate": True,
-                        "unlocked": True,
-                        "handbook_note": "", # TODO: Cache handbook notes
-                        "warnings": []
-                    }
-                continue
-            
-            # Otherwise, we will evaluate all the courses in the current term.
-            for course in term["courses"]:
-                user.add_current_course(course) # Helpful for corequisite conditions
-
-            # Evaluate the courses
-            for course in term["courses"]:
-                if course not in CONDITIONS:
-                    # Error check in case we do not have the condition item
-                    coursesState[course] = {
-                        "is_accurate": False,
-                        "unlocked": True,
-                        "handbook_note": "",
-                        "warnings": []
-                    }
-                    continue
-                    
-                # We do have the course for this condition...
-                condition = CONDITIONS[course]
-                if condition is None:
-                    # Our algorithm could not parse it :(
-                    coursesState[course] = {
-                        "is_accurate": False,
-                        "unlocked": True,
-                        "handbook_note": "", # TODO: Cache handbook notes
-                        "warnings": []
-                    }
-                else:
-                    # Evaluate the condition accordingly
-                    res = condition.is_unlocked(user)
-                    coursesState[course] = {
-                        "is_accurate": True,
-                        "unlocked": res["result"],
-                        "handbook_note": "", # TODO: Cache handbook notes
-                        "warnings": res["warnings"]
-                    }
-            
-            # Add all these courses to the user in preparation for the next term
-            user.empty_current_courses()
-            user.add_courses(term["courses"])
-
-        # Onto the next year...
-        user.year += 1
-
-    return {"courses_state": coursesState}
-
-
-
-
-
