+++ backend/server/server.py
@@ -1,10 +1,9 @@
-import sys
 from fastapi import FastAPI
 from server.routers import specialisations
 from server.routers import programs
 from server.routers import courses
 from server.routers import api
-from bson.json_util import dumps
+from server.routers import planner
 from fastapi.middleware.cors import CORSMiddleware
 
 app = FastAPI()
@@ -27,6 +26,7 @@ app.add_middleware(
 
 app.include_router(specialisations.router)
 app.include_router(programs.router)
+app.include_router(planner.router)
 app.include_router(courses.router)
 app.include_router(api.router)
